claude: refactor(kaspa): rename bridge to ops
The name 'bridge' was too generic since the entire project is about
a bridge. The module contains operation types (DepositFXG, WithdrawFXG,
ConfirmationFXG) - the wire formats exchanged between validators and
relayers. 'ops' (operations) better describes its purpose.
