+++ rust/main/chains/dymension-kaspa/src/kas_validator/confirmation.rs
@@ -1,5 +1,5 @@
-use crate::bridge::confirmation::ConfirmationFXG;
-use crate::bridge::payload::{MessageID, MessageIDs};
+use crate::ops::confirmation::ConfirmationFXG;
+use crate::ops::payload::{MessageID, MessageIDs};
 use crate::kas_validator::error::ValidationError;
 use dym_kas_api::models::{TxModel, TxOutput};
 use dym_kas_core::api::client::HttpClient;
