+++ rust/main/chains/dymension-kaspa/src/kas_validator/withdraw.rs
@@ -1,9 +1,9 @@
 // We call the signers 'validators'
 
-use crate::bridge::payload::MessageIDs;
-use crate::bridge::util;
-use crate::bridge::util::get_recipient_script_pubkey;
-use crate::bridge::withdraw::{filter_pending_withdrawals, WithdrawFXG};
+use crate::ops::payload::MessageIDs;
+use crate::ops::util;
+use crate::ops::util::get_recipient_script_pubkey;
+use crate::ops::withdraw::{filter_pending_withdrawals, WithdrawFXG};
 use crate::kas_validator::error::ValidationError;
 use dym_kas_core::escrow::*;
 use dym_kas_core::pskt::is_valid_sighash_type;
