+++ rust/main/chains/dymension-kaspa/src/ops/confirmation.rs
@@ -1,4 +1,4 @@
-use crate::bridge::payload::MessageID;
+use crate::ops::payload::MessageID;
 use bytes::Bytes;
 use eyre::Error as EyreError;
 use hex::ToHex;
