+++ rust/main/hyperlane-base/src/kas_hack/logic_loop.rs
@@ -1,7 +1,7 @@
 use crate::contract_sync::cursors::Indexable;
 use dym_kas_core::finality::is_safe_against_reorg;
 use dymension_kaspa::hl_message::add_kaspa_metadata_hl_messsage;
-use dymension_kaspa::bridge::{confirmation::ConfirmationFXG, deposit::DepositFXG};
+use dymension_kaspa::ops::{confirmation::ConfirmationFXG, deposit::DepositFXG};
 use dymension_kaspa::kas_relayer::confirm::expensive_trace_transactions;
 use dymension_kaspa::kas_relayer::deposit::{
     build_deposit_fxg, check_deposit_finality, KaspaTxError,
