+++ rust/main/utils/kaspa-tools/src/x/addr.rs
@@ -1,4 +1,4 @@
-use dymension_kaspa::bridge::util::kaspa_address_to_hex_recipient;
+use dymension_kaspa::ops::util::kaspa_address_to_hex_recipient;
 
 /// Convert a kaspa addr (like kaspatest:qr0jmjgh2sx88q9gdegl449cuygp5rh6yarn5h9fh97whprvcsp2ksjkx456f)
 /// to something that can be passed to Hyperlane on the hub in the transfer recipient field (like 0xdf2dc917540c7380a86e51fad4b8e1101a0efa27473a5ca9b97ceb846cc402ab)
@@ -9,7 +9,7 @@ pub fn hl_recipient(kaspa_addr: &str) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use dymension_kaspa::bridge::util::get_recipient_address;
+    use dymension_kaspa::ops::util::get_recipient_address;
     use hyperlane_core::H256;
     use kaspa_addresses::Prefix;
